Fix some warnings and errors on GCC3/linux
git-svn-id: svn://svn.eoserv.net/eoserv/trunk@101 da1c5d5e-72c9-4536-8af9-af1c9134b0e6

+++ src/eoserv/world.cpp
@@ -147,7 +147,7 @@ World::World(util::array<std::string, 5> dbinfo, Config config)
 	exp_table[0] = 0;
 	for (std::size_t i = 1; i < sizeof(this->exp_table)/sizeof(int); ++i)
 	{
-		exp_table[i] = util::round(std::pow(i, 3) * 133.1);
+		exp_table[i] = int(util::round(std::pow(double(i), 3.0) * 133.1));
 	}
 }
